+++ Source/Lib/Encoder/ASM_AVX2/Quant_avx2.c
@@ -10,7 +10,7 @@
 #include <immintrin.h>
 
 static void quant_deadzone_avx2(uint16_t* buf, uint32_t size, int8_t gtli) {
-    uint16_t clamp = ((uint16_t)(~(uint16_t)0) << gtli) & (~BITSTREAM_MASK_SIGN);
+    uint16_t clamp = (uint16_t)(((unsigned)(uint16_t)(~(uint16_t)0) << gtli) & ~(unsigned)BITSTREAM_MASK_SIGN);
     uint32_t coeff_idx = 0;
 
     __m256i zero = _mm256_setzero_si256();
